LayoutTests/fast/scrolling/ios/overflow-scroll-pointer-events.html

<!DOCTYPE html> <!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true ] -->
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/basic-gestures.js"></script>
<style>
.disable-scroll { pointer-events: none; }
.outer-scroll { height: 100%; }
.inner-scroll { height: 100%; overflow-y: auto; }
.box { height: 100%; }
</style>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

async function runTest() {
    if (!window.testRunner) {
        finishJSTest();
        return;
    }

    description("Tests that toggling `pointer-events: none;` doesn't prevent scrolling.");

    globalThis.outerScrollElement = document.querySelector(".outer-scroll");
    globalThis.innerScrollElement = document.querySelector(".inner-scroll");

    shouldBe("innerScrollElement.scrollTop", "0");

    debug("Disabling pointer events...");
    outerScrollElement.classList.add("disable-scroll");
    await UIHelper.animationFrame();
    await touchAndDragFromPointToPoint(150, 300, 150, 150);
    await liftUpAtPoint(150, 150);
    await UIHelper.animationFrame();
    shouldBe("innerScrollElement.scrollTop", "0");

    debug("Enabling pointer events...");
    outerScrollElement.classList.remove("disable-scroll");
    await UIHelper.animationFrame();
    await touchAndDragFromPointToPoint(150, 300, 150, 150);
    await liftUpAtPoint(150, 150);
    await UIHelper.animationFrame();
    shouldBe("innerScrollElement.scrollTop", "150");

    finishJSTest();
}
</script>
<body onload="runTest()">
    <div class="outer-scroll">
        <div class="inner-scroll">
            <div class="box" style="background-color: red;"></div>
            <div class="box" style="background-color: green;"></div>
            <div class="box" style="background-color: blue;"></div>
        </div>
    </div>
    <p id="description"></p>
    <div id="console"></div>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
